+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.5.0] - 2019-01-16
+
 ### Added
 - Field method `canSee` [759](https://github.com/orchidsoftware/platform/pull/759)